+++ lib/azure_blob/const.rb
@@ -4,6 +4,7 @@ module AzureBlob
   API_VERSION = "2024-05-04"
   MAX_UPLOAD_SIZE = 256 * 1024 * 1024 # 256 Megabytes
   DEFAULT_BLOCK_SIZE = 128 * 1024 * 1024 # 128 Megabytes
+  MAX_BLOB_NAME_LENGTH = 1024
   BLOB_SERVICE = "b"
   CLOUD_REGIONS_SUFFIX = {
     global: "core.windows.net",